0x0+ instructions done

+++ include/instructions.hpp
@@ -15,7 +15,7 @@ class Instructions final : public Singleton<Instructions> {
 	public:
 		static constexpr const uint8_t PREFIX_BYTE = 0xCB;
 
-		using Instruction = std::function<void(uint8_t&, Registers&, MemoryBus&)>;
+		using Instruction = std::function<void(unsigned int&, Registers&, MemoryBus&)>;
 
 		Instructions();
